requirements.txt
asgiref
certifi
chardet
defusedxml
dj-database-url
Django==3.2.9
django-allauth
django-ckeditor
django-js-asset
django-multiselectfield
gunicorn
idna
oauthlib
Pillow
psycopg2-binary
python3-openid
pytz
requests
requests-oauthlib
sqlparse
urllib3
whitenoise
certifi
chardet
defusedxml
dj-database-url
Django
django-allauth
django-ckeditor
django-js-asset
django-multiselectfield
gunicorn
idna
oauthlib
Pillow
psycopg2-binary
python3-openid
pytz
requests
requests-oauthlib
sqlparse
urllib3
whitenoise
gunicorn